pyproject.toml
[tool.poetry]
name = "flaskc"
version = "0.1.0"
description = ""
authors = ["Marcus Pereira <oi@space.dev.br>"]

[tool.poetry.dependencies]
python = "^3.8"
flask-restful = "0.3.8"
flask-sqlalchemy = "2.4.4"
flask-migrate = "2.5.3"
flask-cors = "3.0.9"
picpay = "1.1.1"
dynaconf = "3.1.0"
psycopg2-binary = "2.8.6"
flask-jwt-extended = "3.25.0"
flask-admin = "1.5.6"
python-slugify = "4.0.1"
flask-mail = "0.9.1"

[tool.poetry.dev-dependencies]
python-dotenv = "0.14.0"
pytest = "6.0.2"
pytest-flask = "1.0.0"
pytest-regressions = "2.0.1"
black = "18.5b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"